+++ Model/essBuildInc/ButterflyModerator.h
@@ -75,7 +75,7 @@ class ButterflyModerator :
   virtual ~ButterflyModerator();
   
   virtual Geometry::Vec3D getSurfacePoint(const size_t,const long int) const;
-  virtual int getLayerSurf(const size_t,const size_t) const;
+  virtual int getLayerSurf(const size_t,const long int) const;
   virtual std::string getLayerString(const size_t,const size_t) const;
   virtual int getCommonSurf(const size_t) const;
