+++ Model/t1UpgradeInc/CH4Layer.h
@@ -78,7 +78,7 @@ class CH4Layer : public constructSystem::ModBase
   virtual Geometry::Vec3D 
     getSurfacePoint(const size_t,const long int) const;
   virtual int 
-    getLayerSurf(const size_t,const size_t) const;
+    getLayerSurf(const size_t,const long int) const;
   virtual std::string 
     getLayerString(const size_t,const size_t) const;
